requirements.txt
faiss-cpu
numpy
sentence-transformers
google-generativeai
python-dotenv
torch
transformers
Flask
gunicorn
flask_cors
cachetools
gTTS
SpeechRecognition
